design : latout - Header, RecentlyViewedBooks component
디자인 수정

Related to : #130
