+++ src/styles/layout/header/search.module.css
@@ -1,18 +1,19 @@
 .searchForm {
-	width: 400px;
-	height: 60px;
+	width: inherit;
+	height: 50px;
 
 	display: flex;
 	align-items: center;
 	position: relative;
 }
 .searchForm > input {
-	width: 98%;
+	width: 100%;
 	height: 60%;
 	border-radius: 50px;
 	padding: 0 20px;
 	font-size: 11px;
 
+	box-sizing: border-box;
 	border: none;
 	outline: none;
