+++ src/styles/layout/recentlyViewedBooks.module.css
@@ -53,10 +53,10 @@
 }
 
 .deleteButton {
-	width: 20px;
-	height: 20px;
+	width: 16px;
+	height: 16px;
 	top: 0px;
-	right: 0px;
+	right: -16px;
 
 	position: absolute;
